fix(cluster): bot-review fixes — RO gate memory-tier forms, connect-failure accounting, UUID retry, 64-bit stat binds
- Admin_Handler.cpp: extend the RO-gate refusal in
  admin_handler_command_load_or_save() to LOAD ... FROM DISK/FROM CONFIG,
  LOAD/SAVE ... TO MEMORY/TO MEM, and SAVE ... FROM RUNTIME/FROM RUN. These
  mutate the memory tier via C++ flush functions that bypass PRAGMA
  query_only, so they need the same follower refusal as the TO RUNTIME/TO
  DISK/FROM MEMORY forms already handled. Also replace an strlen() on a
  null-checked GLOBAL_UUID() value with strnlen(x, 64).
- ProxySQL_Cluster.cpp: record a checks_ERR via Update_Node_Failure() on
  both the mysql_real_connect() failure path and the initial "SELECT
  @@version" failure path of the cluster monitor thread — previously only
  GLOBAL_CHECKSUM() failures were counted, undercounting real connectivity
  problems. Add a uuid_known latch so a failed/short GLOBAL_UUID() fetch is
  retried on subsequent successful poll iterations instead of only once
  right after the version handshake, without re-querying once known.
  Replace strlen() with strnlen(x, 64) on the peer-supplied UUID.
- ProxySQL_Admin_Stats.cpp: stats___proxysql_servers_status() binds weight,
  global_version, checks_OK and checks_ERR with atoll() instead of atoi()
  to avoid truncating 64-bit counters; port stays atoi().
- test_cluster_leader_election-t.cpp: add follower-refusal assertions for
  LOAD MYSQL SERVERS FROM DISK and SAVE MYSQL SERVERS FROM RUNTIME
  (plan 30 -> 32).

+++ lib/ProxySQL_Admin_Stats.cpp
@@ -1604,13 +1604,13 @@ void ProxySQL_Admin::stats___proxysql_servers_status() {
 			SQLite3_row *r1=*it;
 			rc=(*proxy_sqlite3_bind_text)(statement1, 1, r1->fields[0], -1, SQLITE_TRANSIENT); ASSERT_SQLITE_OK(rc, statsdb);
 			rc=(*proxy_sqlite3_bind_int64)(statement1, 2, atoi(r1->fields[1])); ASSERT_SQLITE_OK(rc, statsdb);
-			rc=(*proxy_sqlite3_bind_int64)(statement1, 3, atoi(r1->fields[2])); ASSERT_SQLITE_OK(rc, statsdb);
+			rc=(*proxy_sqlite3_bind_int64)(statement1, 3, atoll(r1->fields[2])); ASSERT_SQLITE_OK(rc, statsdb);
 			rc=(*proxy_sqlite3_bind_text)(statement1, 4, r1->fields[3], -1, SQLITE_TRANSIENT); ASSERT_SQLITE_OK(rc, statsdb);
-			rc=(*proxy_sqlite3_bind_int64)(statement1, 5, atoi(r1->fields[4])); ASSERT_SQLITE_OK(rc, statsdb);
+			rc=(*proxy_sqlite3_bind_int64)(statement1, 5, atoll(r1->fields[4])); ASSERT_SQLITE_OK(rc, statsdb);
 			rc=(*proxy_sqlite3_bind_int64)(statement1, 6, atoll(r1->fields[5])); ASSERT_SQLITE_OK(rc, statsdb);
 			rc=(*proxy_sqlite3_bind_int64)(statement1, 7, atoll(r1->fields[6])); ASSERT_SQLITE_OK(rc, statsdb);
-			rc=(*proxy_sqlite3_bind_int64)(statement1, 8, atoi(r1->fields[7])); ASSERT_SQLITE_OK(rc, statsdb);
-			rc=(*proxy_sqlite3_bind_int64)(statement1, 9, atoi(r1->fields[8])); ASSERT_SQLITE_OK(rc, statsdb);
+			rc=(*proxy_sqlite3_bind_int64)(statement1, 8, atoll(r1->fields[7])); ASSERT_SQLITE_OK(rc, statsdb);
+			rc=(*proxy_sqlite3_bind_int64)(statement1, 9, atoll(r1->fields[8])); ASSERT_SQLITE_OK(rc, statsdb);
 			rc=(*proxy_sqlite3_bind_text)(statement1, 10, r1->fields[9], -1, SQLITE_TRANSIENT); ASSERT_SQLITE_OK(rc, statsdb);
 			SAFE_SQLITE3_STEP2(statement1);
 			rc=(*proxy_sqlite3_clear_bindings)(statement1); ASSERT_SQLITE_OK(rc, statsdb);
